+++ packages/moonstone/Panels/Header.js
@@ -6,13 +6,24 @@ import {Layout, Cell} from '@enact/ui/Layout';
 import Slottable from '@enact/ui/Slottable';
 import ComponentOverride from '@enact/ui/ComponentOverride';
 
-import {MarqueeDecorator} from '../Marquee';
+import {MarqueeDecorator, MarqueeBase} from '../Marquee';
 import Skinnable from '../Skinnable';
 
 import css from './Header.module.less';
 
+// Extract the spacing class to override the title Marquee instance
+const marqueeCss = {
+	spacing: css.spacing
+};
+
+const TitleMarquee = (props) => {
+	return (
+		<MarqueeBase {...props} css={marqueeCss} />
+	);
+};
+
 // Create a <h1> and Marquee component
-const MarqueeH1 = MarqueeDecorator('h1');
+const MarqueeH1 = MarqueeDecorator({component: TitleMarquee}, 'h1');
 const MarqueeH2 = MarqueeDecorator('h2');
 
 const CompactTitleBase = kind({
@@ -206,8 +217,10 @@ const HeaderBase = kind({
 				);
 			} else {
 				return (
-					<Cell component={MarqueeH1} className={css.title} marqueeOn={marqueeOn} alignment={centered ? 'center' : null}>
-						{title}
+					<Cell>
+						<MarqueeH1 className={css.title} css={marqueeCss} marqueeOn={marqueeOn} alignment={centered ? 'center' : null}>
+							{title}
+						</MarqueeH1>
 					</Cell>
 				);
 			}