+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [3.2.0] - 2019-10-18
+
+### Added
+
+- `ui/Marquee` prop `marqueeSpacing` and CSS class `spacing` to configure the spacing between the repeated content
+
+### Changed
+
+- `ui/Marquee` to have a wrap-around effect
+
+### Fixed
+
+- `ui/Marquee` to start on focus when disabled
+
 ## [3.1.3] - 2019-10-09
 
 ### Fixed